+++ test/doc_insert_test.dart
@@ -20,7 +20,7 @@ void main() {
           nextSelection: TextSelection(baseOffset: 1, extentOffset: 1));
 
       //--- when
-      doc.updateText(deltaText, {Formatus.paragraph}, null);
+      doc.updateText(deltaText, {Formatus.paragraph}, '');
 
       //--- then
       expect(doc.textNodes.length, 1);
@@ -43,7 +43,7 @@ void main() {
           nextSelection: TextSelection(baseOffset: 1, extentOffset: 1));
 
       //--- when
-      doc.updateText(deltaText, {Formatus.paragraph}, null);
+      doc.updateText(deltaText, {Formatus.paragraph}, '');
 
       //--- then
       expect(doc.textNodes.length, 10);
@@ -66,7 +66,7 @@ void main() {
           nextSelection: TextSelection(baseOffset: 4, extentOffset: 4));
 
       //--- when
-      doc.updateText(deltaText, {Formatus.header1}, null);
+      doc.updateText(deltaText, {Formatus.header1}, '');
 
       //--- then
       expect(doc.textNodes.length, 3);
@@ -89,7 +89,7 @@ void main() {
           nextSelection: TextSelection(baseOffset: 8, extentOffset: 8));
 
       //--- when
-      doc.updateText(deltaText, {Formatus.paragraph}, null);
+      doc.updateText(deltaText, {Formatus.paragraph}, '');
 
       //--- then
       expect(doc.textNodes.length, 3);
@@ -112,7 +112,7 @@ void main() {
           nextSelection: TextSelection(baseOffset: 3, extentOffset: 3));
 
       //--- when
-      doc.updateText(deltaText, {Formatus.header1}, null);
+      doc.updateText(deltaText, {Formatus.header1}, '');
 
       //--- then
       expect(doc.textNodes.length, 3);
@@ -135,7 +135,7 @@ void main() {
           nextSelection: TextSelection(baseOffset: 7, extentOffset: 7));
 
       //--- when
-      doc.updateText(deltaText, {Formatus.header1, Formatus.bold}, null);
+      doc.updateText(deltaText, {Formatus.header1, Formatus.bold}, '');
 
       //--- then
       expect(doc.textNodes.length, 4);
@@ -161,7 +161,7 @@ void main() {
           nextSelection: TextSelection(baseOffset: 7, extentOffset: 7));
 
       //--- when
-      doc.updateText(deltaText, {Formatus.header1, Formatus.underline}, null);
+      doc.updateText(deltaText, {Formatus.header1, Formatus.underline}, '');
 
       //--- then
       expect(doc.textNodes.length, 6);